bogatyirev.lab_2

+++ app/G9/bogatyirev.kirill/lab2/src/main/java/ru/vsu/amm/java/interfaces/Circlable.java
@@ -1,5 +1,5 @@
 package ru.vsu.amm.java.interfaces;
 
 public interface Circlable {
-    public int findRadius();
+    public double findRadius();
 }
